akses, middlewere di api, dan status di reservasi resource

+++ app/Http/Controllers/Kelas/KelasController.php
@@ -6,8 +6,8 @@ use App\Models\Kelas;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Resources\Kelas\KelasResource;
-use App\Http\Resources\Jadwal\JadwalResource;
 
 class KelasController extends Controller
 {
@@ -16,9 +16,13 @@ class KelasController extends Controller
      */
     public function index()
     {
-        $kelas = DB::table('kelas')->get();
-        return KelasResource::collection($kelas);
-        return $kelas;
+         if (Auth::user()->role == 1)
+        {
+            $kelas = DB::table('kelas')->get();
+            return KelasResource::collection($kelas);
+            return $kelas;
+        }
+        return['Anda tidak memiliki akses'];
     }
 
 
@@ -28,17 +32,21 @@ class KelasController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'nama'      => 'required',
-            'semester'  => 'required'
+        if (Auth::user()->role == 1)
+        {
+            $request->validate([
+                'nama'      => 'required',
+                'semester'  => 'required'
 
-        ]);
-        $add = Kelas::create([
-            'nama'      => $request->nama,
-            'semester'  => $request->semester,
+            ]);
+            $add = Kelas::create([
+                'nama'      => $request->nama,
+                'semester'  => $request->semester,
 
-        ]);
-        return $add;
+            ]);
+            return $add;
+        }
+        return['Anda tidak memiliki akses'];
     }
 
 
@@ -48,14 +56,16 @@ class KelasController extends Controller
      * Display the specified resource.
      */
     public function show($id)
-    {
-        $kelasexist = Kelas::where('id', '=', $id)->first();
-        if ($kelasexist)
+    {   if (Auth::user()->role == 1)
+        {
+            $kelasexist = Kelas::where('id', '=', $id)->first();
+            if ($kelasexist)
             {
-            return new KelasResource($kelasexist);
+                return new KelasResource($kelasexist);
             }
-
-            return ['Kelas tidak ditemukan'];
+                return ['Kelas tidak ditemukan'];
+        }
+        return['Anda tidak memiliki akses'];
     }
 
 
@@ -65,17 +75,20 @@ class KelasController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $kelasexist = Kelas::where('id', '=', $id)->first();
-        if ($kelasexist) {
-            $kelasexist->update([
-                'nama' => $request->nama ?? $kelasexist->name,
-                'semester' => $request->semester ?? $kelasexist->semester,
-            ]);
-            return ['berhasil di update', $kelasexist];
+        if (Auth::user()->role == 1)
+        {
+            $kelasexist = Kelas::where('id', '=', $id)->first();
+            if ($kelasexist) {
+                $kelasexist->update([
+                    'nama' => $request->nama ?? $kelasexist->name,
+                    'semester' => $request->semester ?? $kelasexist->semester,
+                ]);
+                return ['berhasil di update', $kelasexist];
 
+            }
+                return ['Kelas tidak ditemukan'];
         }
-
-            return ['Kelas tidak ditemukan'];
+        return['Anda tidak memiliki akses'];
     }
 
 
@@ -85,11 +98,15 @@ class KelasController extends Controller
      */
     public function destroy(string $id)
     {
-        $kelasexist = Kelas::where('id', '=', $id)->first();
-        if ($kelasexist) {
-            $kelasexist->delete();
-            return ['kelas berhasil di hapus'];
+        if (Auth::user()->role == 1)
+        {
+            $kelasexist = Kelas::where('id', '=', $id)->first();
+            if ($kelasexist) {
+                $kelasexist->delete();
+                return ['kelas berhasil di hapus'];
+            }
+            return ['kelas tidak ditemukan'];
         }
-        return ['kelas tidak ditemukan'];
+        return['Anda tidak memiliki akses'];
     }
 }
